bug fixes in actions

+++ src/json/map/tiled.rs
@@ -133,7 +133,7 @@ impl TiledMap {
         let mut tilesets = HashMap::new();
         for tiled_tileset in self.tilesets {
             let texture_size = uvec2(tiled_tileset.imagewidth as u32, tiled_tileset.imageheight as u32);
-            let tile_size = uvec2(tiled_tileset.tilewidth as u32, tiled_tileset.tileheight as u32);
+            let tile_size = vec2(tiled_tileset.tilewidth as f32, tiled_tileset.tileheight as f32);
             let grid_size = uvec2(tiled_tileset.columns as u32, tiled_tileset.tilecount as u32 / tiled_tileset.columns as u32);
 
             let mut tile_attributes = HashMap::new();
